Move more items into connectivity module
